Makes the test respond to the correct error.
With an empty array as the first argument, the test will pass if the
student were to throw an error for empty arrays (a different test case).

Using `products` as an argument, as in the other test cases, isolates
the test to _only_ pass if an error is thrown for this particular case.